+++ Cindy3D/src/de/tum/in/cindy3dplugin/jogl/primitives/Point.java
@@ -9,7 +9,7 @@ public class Point extends Primitive {
 	public double size;
 
 	public Point(double x, double y, double z, double size, Color color,
-			double shininess, double alpha) {
+			int shininess, double alpha) {
 		super(color, shininess, alpha);
 		this.x = x;
 		this.y = y;
